Исправила несколько классов в один
